requirement.txt
deepgram-sdk
requests
python-dotenv
groq
langchain_groq
langchain-community
PyAudio
streamlit
